requirements/base.txt
alabaster==0.7.16
amqp==5.2.0
aniso8601==9.0.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.1
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cryptography==42.0.7
cssbeautifier==1.15.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.13
django-allauth==0.61.1
django-anymail==10.2
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.4.0
django-redis==5.4.0
django-simple-history==3.7.0
django-stubs==4.2.7
django-stubs-ext==5.0.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
djlint==1.34.1
docutils==0.20.1
drf-spectacular==0.27.1
easy-thumbnails==2.8.5
EditorConfig==0.12.4
executing==2.0.1
factory-boy==3.3.0
Faker==25.3.0
filelock==3.14.0
flower==2.0.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
hiredis==2.3.2
html-tag-names==0.1.2
html-void-elements==0.1.0
humanize==4.9.0
identify==2.5.36
idna==3.7
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jq==1.7.0
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
livereload==2.6.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.9.0
oauthlib==3.2.2
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
playwright==1.44.0
pluggy==1.5.0
pre-commit==3.6.2
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.45
psycopg==3.1.18
psycopg-c==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.0.1
pytest-django==4.8.0
pytest-sugar==1.0.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
ruff==0.2.2
sentry-sdk==1.40.5
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sorl-thumbnail==12.10.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-autobuild==2024.2.4
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
stack-data==0.6.3
termcolor==2.4.0
text-unidecode==1.3
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240602
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
